pyproject.toml
[project]
name = "kafori"
version = "0.1.0"
description = ""
authors = [
    {name = "Feel Liao",email = "feelteel@hotmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "snakemake (>=8.29.2,<10.0.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "rpy2 (>=3.6.1,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "pytest-cov (>=6.2.1,<7.0.0)",
    "fastparquet (>=2024.11.0,<2025.0.0)",
    "tortoise-orm (==0.25.1)",
    "redis (==6.2.0)",
    "aiomysql (>=0.2.0,<0.3.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = "tests"
python_files = "tests.py test_*.py *_tests.py"

[tool.pylint.design]
max-line-length = 120

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
